+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## Features/Changes
 * Compiler/wasm: omit code pointer from closures when not used (#2059)
 * Compiler: fix purity of comparison functions (again) (#2092)
+* Compiler/wasm: unbox numbers within functions (#2069)
 
 # 6.2.0 (2025-07-30) - Lille